% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_and_process.R
\name{export_tree_list}
\alias{export_tree_list}
\title{Title Export initial SORTIE plot}
\usage{
export_tree_list(tree_lists)
}
\arguments{
\item{tree_lists}{Initial SORTIE starting conditions plot exported from tree_list_sortie}
}
\description{
This function exports initial SORTIE starting conditions as a CSV
}
