% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{import.psp}
\alias{import.psp}
\title{Import PSP data}
\usage{
import.psp(r.path, dat.type, tsas)
}
\arguments{
\item{r.path}{pathway to data}

\item{dat.type}{samples or trees?}

\item{tsas}{which tsas to include (vector of characters I think)}
}
\description{
Import PSP data
}
\details{
Import tree and samples data
}
