% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_and_select.R
\name{import_trees}
\alias{import_trees}
\title{Title Import PSP Tree Data}
\usage{
import_trees(data_path, tsas, selected_plots)
}
\arguments{
\item{data_path}{Path to tree data}

\item{tsas}{TSAs to select tree data for}

\item{selected_plots}{Plot names selected in sel.psp function}
}
\description{
This function imports PSP tree data corresponding to selected sample plots
}
