% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{importPSP}
\alias{importPSP}
\title{Import PSP data}
\usage{
importPSP(data_path, data_type, tsas)
}
\arguments{
\item{data_path}{pathway to data}

\item{data_type}{samples or trees?}

\item{tsas}{which tsas to include (vector of characters I think)}
}
\description{
Import PSP data
}
\details{
Import tree and samples data
}
