% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_and_process.R
\name{psp_remeas_years}
\alias{psp_remeas_years}
\title{Title Calculate number of years and age for PSP SORTIE run}
\usage{
psp_remeas_years(plot.SORTIE, tree.dat, samples.dt, age.crit)
}
\arguments{
\item{plot.SORTIE}{Selected plots to be included}

\item{tree.dat}{Tree data imported for selected plots}

\item{samples.dt}{Imported sample data}

\item{age.crit}{}
}
\description{
Title Calculate number of years and age for PSP SORTIE run
}
