% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_and_process.R
\name{psp_tree_meas}
\alias{psp_tree_meas}
\title{Title Prepare Tree Data for Import}
\usage{
psp_tree_meas(data_path, tsas, selected_plots)
}
\arguments{
\item{data_path}{Path to tree data}

\item{tsas}{TSAs to select tree data for}

\item{selected_plots}{Plot names selected in sel.psp function}
}
\description{
we consider live trees here to be:L=live, I=Live ingrowth exceeding minimum DBH tagging
limit (if a subplot tree) or exceeding breakpoint DBH limit (if in main plot),
and V=live vet. Dead is everything else (DP=dead potential (<50\% decay),
DU=dead useless (>=50\% decay), C=cut, X=ingrowth less than minimum DBH of 4.0cm)
}
