% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_and_process.R
\name{tree_list_sortie}
\alias{tree_list_sortie}
\title{Title Create initial tree population table for SORTIE model}
\usage{
tree_list_sortie(sizeClasses, SORTIE.species, PSP.measurements)
}
\arguments{
\item{PSP.measurements}{}

\item{size_classes}{Vector of desired size classes}
}
\description{
Title Create initial tree population table for SORTIE model
}
