% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_and_select.R
\name{import_psps}
\alias{import_psps}
\title{Title Import PSP Sample Data}
\usage{
import_psps(data_path, data_type = "Trees", tsas = TRUE)
}
\arguments{
\item{data_path}{Path to PSP sample data}

\item{data_type}{is it trees or samples. default is trees}

\item{tsas}{TSA(s) to be imported (vector of characters). default is import all}
}
\description{
This function imports PSP sample data
}
\details{
Just pass the TSA number if want to overide importing all the data
}
