% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_and_process.R
\name{psp_tree_meas}
\alias{psp_tree_meas}
\title{Title Prepare Tree Data for Import}
\usage{
psp_tree_meas(data_path, tsas, selected_plots)
}
\arguments{
\item{data_path}{Path to tree data}

\item{tsas}{TSAs to select tree data for}

\item{selected_plots}{Plot names selected in sel.psp function}
}
\description{
Title Prepare Tree Data for Import
}
