% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_and_select.R
\name{select_psps}
\alias{select_psps}
\title{Title Select and Clean PSP Sample Data}
\usage{
select_psps(samples_data, BECzone, BEClabel, site_series, min_remeasure)
}
\arguments{
\item{samples_data}{Imported sample data output by importPSP function}

\item{BECzone}{BEC zone(s) to be selected}

\item{BEClabel}{BEC label(s) to be selected}

\item{site_series}{Site series of interest to be selected}

\item{min_remeasure}{Minimum remeasurement interval for selected plots}
}
\description{
This function selects PSP sample plots based on criteria and cleans data
}
